fix: fix follow button bug
